Switch cmd argument of ioctl to u_long as elsewhere in the kernel. Propagate this change down the callchain.
Approved by:	kan (maintainer)
Approved by:	ed (mentor)

+++ sys/gnu/fs/xfs/xfs_vnodeops.c
@@ -4506,7 +4506,7 @@ xfs_free_file_space(
 int
 xfs_change_file_space(
 	bhv_desc_t	*bdp,
-	int		cmd,
+	u_long		cmd,
 	xfs_flock64_t	*bf,
 	xfs_off_t	offset,
 	cred_t		*credp,
